refactor: rename NLPGenHub to projects and migrate IntelliQA from ChromaDB to PGVector
