+++ portfolio/urls.py
@@ -16,20 +16,17 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 from mysite import mysite_views
-from NLPGenHub import NLPGenHub_views
-from mysite import mysite_views
-from NLPGenHub import NLPGenHub_views
+from projects import project_views
 from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('',mysite_views.index, name='index'),
-    path('',mysite_views.index, name='index'),
     
     # Include the app’s URL patterns
     path('mysite/', include('mysite.mysite_urls')),  
-    path('projects/', include('NLPGenHub.NLPGenHub_urls')),  
+    path('projects/', include('projects.project_urls')),  
 ]
 
 if settings.DEBUG:
