+++ projects/management/commands/run_cleanup.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from NLPGenHub.cron_job import cleanup_task
+from projects.cron_job import cleanup_task
 
 class Command(BaseCommand):
     help = 'Runs the cleanup task'
